get GCC version, some further heuristic
It's really only a heuristic.
This is only used if there is no matching major version,
where we just gave up before.
So it might potentially only improve things.

In my particular case, I had GCC 7, 8, 9 installed,
and TF was compiled with GCC 5, and used CUDA 10.0.
This is the official TF 1.14 from pip.
RETURNN picked GCC 9, which did not work together with the nvcc from CUDA 10.0,
which required GCC <=7.
So this heuristic fixes this particular case.
At least it compiles now, and all tests (from test_TFNative) seem to pass.