SelfAttention: Use separate prev key / value when in loop (#348)
This reduces memory footprint when in loop.

When inside a recurrent loop, store a single prev_k and prev_v instead of a concatenation
prev_kv. This saves a split-operation of the prev_kv-tensor with length (t-1) at time step t
into separate prev_k and prev_v, thus having a quadratic memory footprint in t.